requirements.txt
# ═══════════════════════════════════════════════════
# JARVIS ULTIMATE 2026 — все зависимости
# Установка: pip install -r requirements.txt
# ═══════════════════════════════════════════════════

# Telegram
telethon>=1.36.0

# LLM
litellm>=1.48.0

# RAG / База знаний
langchain-core>=0.3.0
langchain-chroma>=0.1.4
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0
chromadb>=0.5.0

# Поиск
ddgs>=0.1.0

# Браузер
playwright>=1.44.0

# Документы
PyPDF2>=3.0.0
python-docx>=1.1.0

# Консоль

# Логи
loguru>=0.7.0

# .env
python-dotenv>=1.0.0

# HTTP
httpx>=0.27.0

# Git (необязательно)
gitpython>=3.1.40
